+++ src/RefDocGen/CodeElements/TypeRegistry/TypeRegistry.cs
@@ -151,76 +151,6 @@ internal class TypeRegistry : ITypeRegistry
         return result;
     }
 
-    internal IReadOnlyList<MemberData> GetDeclaredBaseMembers(MemberData member)
-    {
-        var type = member.ContainingType;
-        var parentTypes = new List<ITypeNameData>();
-        var result = new List<MemberData>();
-
-        var baseType = type.BaseType;
-
-        if (baseType is not null)
-        {
-            parentTypes.Add(baseType);
-        }
-
-        parentTypes.AddRange(type.Interfaces);
-
-        foreach (var parentType in parentTypes)
-        {
-            string parentId = parentType.TypeDeclarationId;
-
-            // the parent type is contained in the type registry
-            if (GetDeclaredType(parentId) is TypeDeclaration parentDecl)
-            {
-                string memberId = member.Id;
-
-                if (member is ICallableMemberData callableMember && callableMember.IsExplicitImplementation)
-                {
-                    string input = callableMember.Id;
-                    int lastIndex = input.LastIndexOf('#');
-
-                    memberId = lastIndex >= 0 && lastIndex < input.Length - 1
-                        ? input[(lastIndex + 1)..]
-                        : string.Empty;
-                }
-
-                if (parentDecl.AllMembers.TryGetValue(memberId, out var parentMember)) // TODO: explicit interface implementation & inheriting generic type
-                {
-                    result.Add(parentMember);
-                    continue;
-                }
-
-                var foundMembers = parentDecl.AllMembers.Where(x => GetInstantiatedMemberId(x.Key, parentType) == member.Id).Select(pair => pair.Value);
-
-                if (foundMembers.Any())
-                {
-                    result.Add(foundMembers.First());
-                }
-            }
-        }
-
-        return result;
-    }
-
-    private string GetInstantiatedMemberId(string memberId, ITypeNameData instantiatedGenericType)
-    {
-        string id = memberId;
-        int index = 0;
-
-        foreach (var typeParam in instantiatedGenericType.TypeParameters)
-        {
-            if (!typeParam.IsGenericParameter)
-            {
-                id = id.Replace($"`{index}", typeParam.Id);
-            }
-
-            index++;
-        }
-
-        return id;
-    }
-
     /// <summary>
     /// Get the data describing the assemblies.
     /// </summary>
